requirements.txt
supabase
streamlit
pyrogram
tgcrypto
orjson
stqdm
